refactor: move hook utility tests to separate file (#1080)
